+++ packages/clis/generator/geo-json/map.ts
@@ -11,6 +11,7 @@ import {
 import { mapValues, putIfAbsent } from '@truckermudgeon/base/map';
 import { Preconditions } from '@truckermudgeon/base/precon';
 import { isLabeledPoi } from '@truckermudgeon/map/constants';
+import { toDealerLabel } from '@truckermudgeon/map/labels';
 import type { Polygon, RoadString } from '@truckermudgeon/map/prefabs';
 import {
   toMapPosition,
@@ -1156,35 +1157,6 @@ function poiToFeature(poi: Poi): PoiFeature {
       coordinates: [poi.x, poi.y],
     },
   };
-}
-
-function toDealerLabel(prefabPath: string): string {
-  Preconditions.checkArgument(prefabPath.includes('/truck_dealer/'));
-  const dealerRegex = /\/truck_dealer\/(?:truck_dealer_([^.]+).ppd$|([^/]+)\/)/;
-  const matches = assertExists(dealerRegex.exec(prefabPath));
-  const dealer = assertExists(matches[1] ?? matches[2]);
-
-  switch (dealer) {
-    case 'mb':
-      return 'Mercedes-Benz';
-    case 'westernstar':
-      return 'Western Star';
-    case 'daf':
-    case 'man':
-      return dealer.toUpperCase();
-    case 'freightliner':
-    case 'international':
-    case 'iveco':
-    case 'kenworth':
-    case 'mack':
-    case 'peterbilt':
-    case 'renault':
-    case 'scania':
-    case 'volvo':
-      return dealer.charAt(0).toUpperCase() + dealer.slice(1);
-    default:
-      throw new Error('unknown dealer: ' + dealer);
-  }
 }
 
 type CityWithScaleRank = City & {
